settings: use config option for netaddress validation
